[refactor] Export the ConnectionPool class
* clean up the test.js
* improve the package.json script
