+++ test.js
@@ -1,62 +1,9 @@
-const PSemaphore = require('promise-semaphore');
+const debug = require('debug')('test');
+
+const ConnectionPool = require('./connection-pool.js');
 const ldap = require('ldapjs');
 const executeQueryPromise = require('./ldapjs-promise');
 
-class ConnectionPool {
-
-    constructor(n, create, destroy) {
-        this.roomsLength = n;
-        this.pSemaphore = new PSemaphore({rooms: this.roomsLength});
-        this.create = create;
-        this.destroy = destroy;
-        this.rooms = [
-            //{ free: false, client: null}
-        ];
-    }
-
-    add(f) {
-        return this.pSemaphore.add(() => {
-            let client = this._get();
-            try {
-                return f(client);
-            } finally {
-                this._put(client);
-            }
-        });
-    }
-
-    _get() {
-        for(let i=0; i < this.roomsLength; i++) {
-            if (!this.rooms[i]) {
-                this.rooms[i] = {
-                    free: true,
-                    client: this.create()
-                }
-            }
-            // Room is furnished
-            if (this.rooms[i].free) {
-                this.rooms[i].free = false;
-                return this.rooms[i].client;
-            } else {
-                // continue
-            }
-        }
-    }
-
-    _put(client) {
-        this.rooms.forEach(room => {
-            if (room.client === client) {
-                room.free = true;
-            }
-        })
-    }
-
-    close() {
-        this.rooms.forEach(room => {
-            this.destroy(room.client);
-        })
-    }
-}
 
 const pool = new ConnectionPool(
     3,
@@ -80,7 +27,8 @@ doSomething = async(client, mail) =>
         filter: `(&(mail=${mail}@epfl.ch))`,
         scope: 'sub',
     };
-    await sleep(1);
+    await sleep(4);
+    debug("Slept...");
     let result = await executeQueryPromise(client, "o=epfl,c=ch", opts);
     console.log(result);
 }
